requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pdf2image==1.17.0
pillow==11.1.0
pygame==2.6.1
PyPDF2==3.0.1
reportlab==4.3.1
Werkzeug==3.1.3
